modularize pad op
