+++ tfjs-core/src/ops/ops.ts
@@ -23,6 +23,11 @@ export {divNoNan} from './div_no_nan';
 export {eye} from './eye';
 export {multinomial} from './multinomial';
 export {oneHot} from './one_hot';
+export {pad} from './pad';
+export {pad1d} from './pad1d';
+export {pad2d} from './pad2d';
+export {pad3d} from './pad3d';
+export {pad4d} from './pad4d';
 export {rand} from './rand';
 export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
