+++ tfjs-core/src/ops/pad.ts
@@ -1,0 +1,69 @@
+/**
+ * @license
+ * Copyright 2020 Google Inc. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {ENGINE, ForwardFunc} from '../engine';
+import {PadV2, PadV2Attrs, PadV2Inputs} from '../kernel_names';
+import {NamedAttrMap} from '../kernel_registry';
+import {Tensor} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+import {op} from './operation';
+
+/**
+ * Pads a `tf.Tensor` with a given value and paddings.
+ *
+ * This operation currently only implements the `CONSTANT` mode.
+ *
+ * Also available are stricter rank-specific methods with the same signature
+ * as this method that assert that `paddings` is of given length.
+ *   - `tf.pad1d`
+ *   - `tf.pad2d`
+ *   - `tf.pad3d`
+ *   - `tf.pad4d`
+ *
+ * ```js
+ * const x = tf.tensor1d([1, 2, 3, 4]);
+ * x.pad([[1, 2]]).print();
+ * ```
+ * @param x The tensor to pad.
+ * @param paddings An array of length `R` (the rank of the tensor), where
+ * each element is a length-2 tuple of ints `[padBefore, padAfter]`,
+ * specifying how much to pad along each dimension of the tensor.
+ * @param constantValue The pad value to use. Defaults to 0.
+ */
+/** @doc {heading: 'Tensors', subheading: 'Transformations'} */
+function pad_<T extends Tensor>(
+    x: T|TensorLike, paddings: Array<[number, number]>, constantValue = 0): T {
+  const $x = convertToTensor(x, 'x', 'pad');
+  if ($x.rank === 0) {
+    throw new Error('pad(scalar) is not defined. Pass non-scalar to pad');
+  }
+  const forward: ForwardFunc<T> = (backend, save) => {
+    save([$x]);
+    return backend.pad($x, paddings, constantValue);
+  };
+
+  const attrs: PadV2Attrs = {paddings, constantValue};
+  const inputs: PadV2Inputs = {x: $x};
+  return ENGINE.runKernelFunc(
+      forward, inputs as unknown as NamedTensorMap, null /* grad */, PadV2,
+      attrs as unknown as NamedAttrMap);
+}
+
+export const pad = op({pad_});
