+++ tfjs-core/src/public/chained_ops/pad.ts
@@ -15,10 +15,18 @@
  * =============================================================================
  */
 
-import './broadcast_to';
-import './div';
-import './div_no_nan';
-import './squared_difference';
-import './tile';
-import './one_hot';
-import './pad';
+import {pad} from '../../ops/pad';
+import {Tensor} from '../../tensor';
+import {Rank} from '../../types';
+
+declare module '../../tensor' {
+  interface Tensor<R extends Rank = Rank> {
+    pad<T extends Tensor>(
+        paddings: Array<[number, number]>, constantValue?: number): T;
+  }
+}
+
+Tensor.prototype.pad = function<T extends Tensor>(
+    this: T, paddings: Array<[number, number]>, constantValue: number): T {
+  return pad(this, paddings, constantValue);
+};
