+++ tfjs-core/src/register_all_gradients.ts
@@ -18,6 +18,7 @@ import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
 import {divGradConfig} from './gradients/Div_grad';
 import {identityGradConfig} from './gradients/Identity_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
+import {padV2GradConfig} from './gradients/PadV2_grad';
 import {squareGradConfig} from './gradients/Square_grad';
 import {squaredDifferenceGradConfig} from './gradients/SquaredDifference_grad';
 import {tileGradConfig} from './gradients/Tile_grad';
@@ -27,7 +28,8 @@ import {registerGradient} from './kernel_registry';
 // Export all kernel configs here so that the package can auto register them
 const gradConfigs: GradConfig[] = [
   divGradConfig, squareGradConfig, squaredDifferenceGradConfig,
-  broadcastToGradConfig, identityGradConfig, tileGradConfig, oneHotGradConfig
+  broadcastToGradConfig, identityGradConfig, tileGradConfig, oneHotGradConfig,
+  padV2GradConfig
 ];
 
 for (const gradientConfig of gradConfigs) {
