+++ tfjs-core/src/tensor.ts
@@ -195,8 +195,6 @@ export interface OpHandler {
   concat<T extends Tensor>(tensors: Array<T|TensorLike>, axis: number): T;
   stack<T extends Tensor>(tensors: Array<T|TensorLike>, axis: number): Tensor;
   unstack<T extends Tensor>(value: T, axis: number): Tensor[];
-  pad<T extends Tensor>(
-      x: T, paddings: Array<[number, number]>, constantValue: number): T;
   batchNorm<R extends Rank>(
       x: Tensor<R>, mean: Tensor<R>|Tensor1D|TensorLike,
       variance: Tensor<R>|Tensor1D|TensorLike,
@@ -819,10 +817,6 @@ export class Tensor<R extends Rank = Rank> {
   }
   unstack(axis = 0): Tensor[] {
     return opHandler.unstack(this, axis);
-  }
-  pad<T extends Tensor>(
-      this: T, paddings: Array<[number, number]>, constantValue = 0): T {
-    return opHandler.pad(this, paddings, constantValue);
   }
   /**
    * @deprecated Use `tf.batchNorm` instead, and note the positional argument
